fix(security): harden VRF and gauge voting epochs
- Finalize gauge checkpoints only for ended epochs (idempotent)
- Unify whitelist eligibility with enumeration; registry gates whitelisting
- Add registry helper, wiring scripts, rollout notes, and tests

+++ contracts/helpers/infra/OFTBootstrapRegistry.sol
@@ -18,7 +18,7 @@ contract OFTBootstrapRegistry {
     /// @notice Return the LayerZero endpoint for any chain.
     /// @dev Always returns LZ_COMMON_ENDPOINT. The chain ID parameter is
     ///      accepted for interface compatibility but has no effect.
-    function getLayerZeroEndpoint(uint16) external pure returns (address) {
+    function getLayerZeroEndpoint(uint256) external pure returns (address) {
         return LZ_COMMON_ENDPOINT;
     }
 }
